refactor: altered vehicles hooks
